Locit: Created new python file _locit_commons.py for functions shared between locit and extract_locit
